Add buttons to refresh page, change cell count
Added a button in html and linked in JavaScript
so that users can click it to refresh the page and clear the drawing.
Added a button in html and a function in JavaScript that asks for a number from the user
and then creates a new grid with the requested number of cells.